requirements.txt
numpy==1.23.5
tensorflow==2.15.0
pillow==10.4.0
flask==2.3.3 # 3.3.3
pandas==2.2.2
# mlflow==2.15.0
matplotlib
h5py==3.11.0